Add stack_pref
